refactor: Type switch provider auth in authproxy config.
